refactor: Update contracts structure
