+++ contracts/ckb-cheque-script/Cargo.toml
@@ -1,11 +1,13 @@
 [package]
 name = "ckb-cheque-script"
 version = "0.1.0"
+authors = ["Dylan <duanyytop@gmail.com>"]
 edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-ckb-std = "0.7.1"
+# For simulator support
+ckb-std = { git = "https://github.com/nervosnetwork/ckb-std", rev = "29455b8" }
 ckb-lib-secp256k1 = { path = "../../ckb-lib-secp256k1" }
 blake2b-ref = "0.1"
