+++ contracts/ckb-cheque-script/src/error.rs
@@ -30,4 +30,3 @@ impl From<SysError> for Error {
         }
     }
 }
-
