+++ contracts/ckb-cheque-script/src/main.rs
@@ -1,7 +1,7 @@
 //! Generated by capsule
 //!
 //! `main.rs` is used to define rust lang items and modules.
-//! See `entry.rs` for the `main` function. 
+//! See `entry.rs` for the `main` function.
 //! See `error.rs` for the `Error` type.
 
 #![no_std]
@@ -13,10 +13,12 @@
 // define modules
 mod entry;
 mod error;
+mod claim;
+mod withdraw;
+mod hash;
+mod helper;
 
-use ckb_std::{
-    default_alloc,
-};
+use ckb_std::default_alloc;
 
 ckb_std::entry!(program_entry);
 default_alloc!();
@@ -29,4 +31,3 @@ fn program_entry() -> i8 {
         Err(err) => err as i8,
     }
 }
-
